Don't create a massive window on hidpi screens
The size I pass for creating the window is now in logical units, so is
creating a gigantic window on a high dpi screen.

So half the size of the window we create, but ensure the old values
are still used for the graphics coordinates.

This will end up with a tiny window on non-hidpi screens, so we need
to work out a better fix.